new file:   .vscode/settings.json 	modified:   GetTSPInput.py 	modified:   JupyterTSP/TSP_CP.ipynb 	new file:   JupyterTSP/__init__.py 	new file:   JupyterTSP/inputTSP.txt 	new file:   __pycache__/GetTSPInput.cpython-311.pyc 	modified:   largeTSP.txt 	modified:   test.py

+++ GetTSPInput.py
@@ -4,5 +4,5 @@ def getInput(file):
 
     with open(file) as data:
         numNode = int(next(data))
-        pathCost = np.loadtxt(data)
+        pathCost = np.loadtxt(data, dtype="int")
     return numNode, pathCost
